AVL TREE CODE
In this lab, we will be experimenting with AVL trees, a type of self-balancing BST tree that
keeps search performance at O(log N). The self-balancing part is a bit of a misnomer, in that
you will have to implement algorithms to achieve tree balance whenever an item/node is
inserted in the AVL tree. Recall from class that these algorithms will include single and
double rotations. Selection of which rotation to use depends on where the new item/node
is inserted in the tree. This may be a good time to pull out your class notes on AVL rotations.
Now to introduce the sensor of the week; for generating our data, we will be using a GY-
273 HMC5883L triple-axis magnetometer in this lab (
Figure 1
). The HMC5883L (the black
chip on the center of the board) is an ultra-low-power (3.6V at 100 microamps), high-
performance 3-axis digital magnetic sensor. It measures magnetic field in all three
directions: X, Y, Z, with the units being milli-Gauss. Magnetometer sensors such as this one
are typically used in smartphones, laptops, robot navigation, and virtual reality device